# tmp_pricelabs_swagger_pretty.json
{
  "swagger": "2.0",
  "info": {
...
    "version": "1.0.0",
    "title": "PriceLabs API Integration",
    "contact": {
      "email": "support@pricelabs.co"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "api.pricelabs.co",
  "basePath": "/v1/integration/api",
...
  "schemes": [
    "https"
  ],
  "security": [
    {
      "integration_token": []
    },
    {
      "integration_name": []
    }
  ],
  "paths": {
...
    "/calendar": {
      "post": {
        "tags": [
          "PMS"
        ],
        "summary": "Update rates and availability",
        "description": "This endpoint can update rates and availability of selected listings, along with the multi unit details of the listings.\n\nlisting_id should always be sent as a string, eg., \"12345\".\n\nThe first time a given date is entered in the data array for a listing it must contain both price, and available_units for the listing. The subsequent API calls can be a delta push or updates for specific date ranges. \n\nIf you need to price only one rate per listing, then you can ignore and skip the field rate_plan_id.\n\nIf you support multiple rate plans and need to update rates for all rate plans, then you should call the /rate_plans endpoint to first update the rate plan IDs and names. Further, you should call the /calendar endpoint mandatorily for default rate plan while providing the respective rate_plan_id of the default rate plan. \nIt is mandatory to set prices for the default rate plan using the /calendar API call.\n\n\n",
        "operationId": "calendar",
        "parameters": [
          {
            "in": "body",
            "name": "calendars",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "calendars": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Calendar"
                  }
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Request was processed, and response contains successes and failures of individually edited listings\nSuccess will contain only an array of listing_ids\nFailure contains an error string describing the problem which halted processing, along with the item it was relevant to",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "description": "Listing id processed without error"
                  }
                },
                "failure": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "details": {
                        "$ref": "#/definitions/Calendar"
                      },
                      "error": {
                        "type": "string",
                        "description": "Error message identifying why processing halted"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid integration API key value",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/integration": {
      "post": {
        "tags": [
          "PMS"
        ],
        "summary": "Update integration information",
        "description": "This endpoint can update integration specific details such as refreshing the API token if needed, or adjusting information that allows Pricelabs to send notices/updates.\n\n<b><i>sync_url</i></b>: Endpoint where Pricelabs can send a pricing data to be synced\n<b><i>calendar_trigger_url</i></b>: Endpoint where PriceLabs requests complete listing calendar information\n<b><i>hook_url</i></b>: Endpoint where PriceLabs can send error notifications.\n\nUpdating these three url endpoints results in a test body being sent with only the json body { verify => true } and no headers and requires a 200 response in return in order to be stored.\n\n When you call <b>/integration</b>, please note that if <b><i>regenerate_token</i></b> is true, you will receive a new token in response which should be used to make subsequent API calls.\n\n<b>Delta Push</b>: If you wish to receive only delta updates for prices and restrictions, then you can do so by setting <b><i>delta_only</i></b> flag as true. Once the flag is set to true, you will start to receive only those prices and restrictions which have changed since the last sync. If at any point, you wish to pull full data, you can do so by calling the /get_prices endpoint with the attribute <b><i>delta_only</i></b> as true.\n\n<b>IMPORTANT</b>: The <i>features</i> array should be used to configure only those settings that the PMS intends to actively read from PriceLabs for all connected listings. \n\n  <i>Please note that if a setting is marked as <b>true</b>, it becomes visible to users in PriceLabs, who may enable it with the expectation that the PMS will consume and apply it. Therefore, make sure to enable only those settings that are actually supported and intended to be consumed from PriceLabs.</i>\n  \n  <b>Once the integration is live, any updates to the supported features must first be discussed with PriceLabs. This allows us to coordinate the release of new features to all users and ensure they function as expected.</b>\n\n\n\n",
        "operationId": "integration",
        "parameters": [
          {
            "in": "body",
            "name": "integration",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "integration": {
                  "$ref": "#/definitions/IntegrationData"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Request was processed and returns a message holding new API if requested and successful, or a simple Success and an error message which is populated on errors. <i>Please note that this response will not contain the signed headers as with the other API call responses.</i>",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Invalid integration API key value",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "integration_token": {
      "type": "apiKey",
      "name": "X-INTEGRATION-TOKEN",
      "in": "header",
      "description": "This is an api token. This is generated and provided by PriceLabs."
    },
    "integration_name": {
      "type": "apiKey",
      "name": "X-INTEGRATION-NAME",
      "in": "header",
      "description": "This is a name field for the pms. The name must be approved and provided by PriceLabs."
    }
  },
  "definitions": {
...
    "Calendar": {
      "type": "object",
      "required": [
        "listing_id",
        "currency",
        "data"
      ],
      "properties": {
        "listing_id": {
          "type": "string",
          "description": "An ID for the listing unique across the PMS system. Should be sent as a string."
        },
        "rate_plan_id": {
          "type": "string",
          "description": "Optional field for rate plan ID to be added if calendar is being updated for a specific rate plan.",
          "example": "IAPI-FLEX-SGL"
        },
        "currency": {
          "type": "string",
          "description": "ISO code of currency of the listing in PMS"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CalendarData"
          }
        },
        "multi_unit": {
          "$ref": "#/definitions/MultiUnitInfo"
        }
      }
    },
    "CalendarData": {
      "type": "object",
      "required": [
        "date",
        "end_date",
        "price",
        "available_units",
        "booked_units",
        "blocked_units"
      ],
      "properties": {
        "date": {
          "type": "string",
          "description": "Start date(current date) in yyyy-mm-dd format"
        },
        "end_date": {
          "type": "string",
          "description": "Optional value. Allows an array input of correct size, AND causes non-array input to be applied to all dates in range. in yyyy-mm-dd format"
        },
        "price": {
          "type": "integer",
          "description": "Price set in the PMS for this date"
        },
        "available_units": {
          "type": "integer",
          "description": "No. of available units of the property on this day"
        },
        "booked_units": {
          "type": "integer",
          "description": "No. of booked units of the property on this day"
        },
        "blocked_units": {
          "type": "integer",
          "description": "No. of blocked units of the property on this day"
        },
        "settings": {
          "$ref": "#/definitions/CalendarSettings"
        }
      }
    },
    "CalendarSettings": {
      "type": "object",
      "description": "You can send default values in these settings, we don't apply these discounts in PL since these are set up directly in PL",
      "properties": {
        "min_stay": {
          "type": "integer",
          "description": "minimum stay for this date"
        },
        "check_in": {
          "type": "boolean",
          "description": "Describes if the property is open for check-in on this day"
        },
        "check_out": {
          "type": "boolean",
          "description": "Describes if the property is open for check-out on this day"
        },
        "weekly_discount": {
          "type": "number",
          "description": "weekly discount if set in PMS"
        },
        "monthly_discount": {
          "type": "number",
          "description": "monthly discount if set in PMS"
        },
        "extra_person_fee": {
          "type": "number",
          "description": "extra guest fees if set in PMS"
        },
        "extra_person_fee_trigger": {
          "type": "number",
          "description": "describes after how many guests the extra person fees applies"
        }
      }
    },
    "MultiUnitInfo": {
      "type": "object",
      "required": [
        "total_units",
        "unit_ids"
      ],
      "properties": {
        "total_units": {
          "type": "integer",
          "description": "The total number of units the listing has.",
          "example": 5
        },
        "unit_ids": {
          "type": "array",
          "example": [
            "101",
            "102",
            "103",
            "104",
            "105"
          ],
          "items": {
            "type": "string",
            "description": "An array of the IDs for the listing. For eg., room number.",
            "example": "101"
          }
        }
      }
    },
    "IntegrationData": {
      "type": "object",
      "required": [
        "sync_url",
        "calendar_trigger_url",
        "hook_url",
        "regenerate_token"
      ],
      "properties": {
        "sync_url": {
          "type": "string"
        },
        "calendar_trigger_url": {
          "type": "string"
        },
        "regenerate_token": {
          "type": "boolean",
          "default": true,
          "description": "If not sent, it will not regenerate token. If sent blank, it will regenerate the token by default."
        },
        "hook_url": {
          "type": "string"
        },
        "features": {
          "$ref": "#/definitions/FeaturesControl"
        }
      }
    },
...
    "FeaturesControl": {
      "type": "object",
      "properties": {
        "min_stay": {
          "type": "boolean"
        },
        "check_in": {
          "type": "boolean"
        },
        "check_out": {
          "type": "boolean"
        },
        "monthly_weekly_discounts": {
          "type": "boolean"
        },
        "extra_person_fee": {
          "type": "boolean"
        },
        "los_pricing": {
          "type": "boolean"
        },
        "delta_only": {
          "type": "boolean",
          "description": "If this flag is set to true, PriceLabs will push only delta to the sync_url. In other words, only those prices and restrictions will be pushed which have changed since the last sync. If this flag is set to false, then all prices and restrictions will be pushed."
        }
      }
    },
...
  }
}